refactor: move yajl_lex to lexer
