+++ crates/yajl/src/lib.rs
@@ -12,12 +12,12 @@
 
 extern crate libc;
 pub(crate) mod buffer;
+pub(crate) mod lexer;
 pub mod parser;
 pub mod status;
 pub mod yajl_alloc;
 pub mod yajl_encode;
 pub mod yajl_gen;
-pub mod yajl_lex;
 pub mod yajl_tree;
 pub mod yajl_version;
